requirements.txt
vosk
sounddevice
pystray
Pillow
dateparser
pytest
openpyxl
symspellpy
